Add files via upload
This is the phase II/III clinical trial data that is constructed from the CANMAT trial. The columns respectively correspond to "EnterTime" (Study entry time in weeks with numbering starting from first recruited patient), "Stage" (1 for stage 1 patients and 0 for stage 2 patients), "Arm" (0 for control, 1 for Treatment 1 which corresponds to mood stabiliser plus 24 weeks of atypical antipsychotic and 2 for Treatment 2 which corresponds to mood stabiliser plus 52 weeks of atypical antipsychotic), "CalendarTime" (At the end of study, survival calendar time in weeks), "T.date" (at the end of the study, survival time in weeks), "death" (at end of study, 1 if relapse and 0 if no relapse), "T.date1" (at the time of interim analysis, survival time in weeks for stage 1 patients) and "death1" (at interim analysis time, 1 if relapse and 0 if no relapse).