Add files via upload
This is the R code used to compute estimates in Section 11.2 in the supplementary material that are based on logrank score statistic. Since the selection rule in the example is different to the one used in Section 2.6 in the main paper, the expression for bias and expected values for the example are given in the supplementary material Section 7.2. These expressions are necessary when computing bias subtracted estimates.